Add Offline Directions tests
