+++ MapboxDirections/Extensions/HTTPURLResponse.swift
@@ -1,0 +1,28 @@
+import Foundation
+
+
+extension HTTPURLResponse {
+    var rateLimit: UInt? {
+        guard let limit = allHeaderFields["X-Rate-Limit-Limit"] as? String else {
+            return nil
+        }
+        return UInt(limit)
+    }
+    
+    var rateLimitInterval: TimeInterval? {
+        guard let interval = allHeaderFields["X-Rate-Limit-Interval"] as? String else {
+            return nil
+        }
+        return TimeInterval(interval)
+    }
+    
+    var rateLimitResetTime: Date? {
+        guard let resetTime = allHeaderFields["X-Rate-Limit-Reset"] as? String else {
+            return nil
+        }
+        guard let resetTimeNumber = Double(resetTime) else {
+            return nil
+        }
+        return Date(timeIntervalSince1970: resetTimeNumber)
+    }
+}
