+++ MapboxDirectionsTests/Fixtures/Offline/versions.json
@@ -1,0 +1,1 @@
+{"availableVersions":["2018-10-16"]}
